requirements.txt
flask==3.0.2
twilio==8.13.0
gspread==5.12.4
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.118.0
gunicorn==21.2.0
python-dotenv==1.0.1 